chore: move biome classes in extra package
